Rename `curry2` to `curry`
I have the feeling that `curry2` would be used 90% of times. How about renaming this to `curry`? @stoeffel @hemanth?
