+++ tests/curry.js
@@ -1,17 +1,17 @@
 import { equal } from 'assert';
-import curry2 from '../curry2';
+import curry from '../curry';
 
-test('#curry2', () => {
+test('#curry', () => {
 	const f = (a, b) => a + b;
-	equal(curry2(f)(1)(2), 3);
+	equal(curry(f)(1)(2), 3);
 
 	const g = (a, b, c) => a + b + c;
 	equal(
-		curry2(g)(1)(2, 3),
+		curry(g)(1)(2, 3),
 		6
 	);
 	equal(
-		curry2(g)(1, 2)(3),
+		curry(g)(1, 2)(3),
 		6
 	);
 });
